mrc-update-expert-location(eplb-move-move,mechanical_provable): Move update_expert_location_with_recovery into eplb_manager.py (cut+paste)
PR-Title: Extract expert location updating into EPLBManager
